+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.17.30-SNAPSHOT
+### is.codion.common.model
+- LoadTest, refactoring and renaming.
 ### is.codion.common.rmi
 - SerializationWhitelist, now allows arrays but checks the component type.
 - SerializationWhitelist.DryRun, now excludes arrays and includes the component type.
@@ -10,6 +12,7 @@ Codion Change Log
 - DomainGeneratorModel, table type column moved.
 - EntityLoadTestModel removed.
 - LoadTest extracted from LoadTestModel and moved to common.model.loadtest, related refactoring.
+- LoadTestModel, refactoring and renaming.
 
 ## 0.17.29
 ### is.codion.common.db